requirements.txt
streamlit
spotipy
SpeechRecognition
streamlit-webrtc
av
python-dotenv